fix: 自定义指标-代码优化 # Reviewed, transaction id: 40213
